Refactored for OOP, fixed indexing bug
