+++ src/test/java/catbot/task/TaskListTest.java
@@ -1,0 +1,5 @@
+package catbot.task;
+
+public class TaskListTest {
+    //todo test indexing for all methods
+}
